requirements.txt
boto3==1.34.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
opensearch-py==2.4.0
pytest==7.4.3
pytest-mock==3.12.0
black==23.12.0
flake8==6.1.0